+++ src/config.ts
@@ -7,7 +7,7 @@ export class Config {
     searchEngine: SearchEngine;
     acEngine: string;
     shouldInsertSelectedText: boolean;
-    linuxDistro?: string;
+    linuxUrlOpener?: string;
     constructor() {
         this.searchEngines = Array<SearchEngine>();
         this.searchEngines.push(
@@ -60,7 +60,7 @@ export class Config {
 
         this.shouldInsertSelectedText = config.get<boolean>("insert_selected_text", this.shouldInsertSelectedText);
 
-        this.linuxDistro = config.get<string>("linux_base_distro");
+        this.linuxUrlOpener = config.get<string>("linux_url_launcher");
     }
 }